Split utils into check_utils
