+++ datadog_checks_dev/datadog_checks/dev/conditions.py
@@ -13,7 +13,7 @@ from six.moves.urllib.request import urlopen
 from .errors import RetryError
 from .structures import LazyFunction
 from .subprocess import run_command
-from .utils import file_exists
+from .fs import file_exists
 
 
 class WaitFor(LazyFunction):
