+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/recommended_monitors.py
@@ -7,7 +7,7 @@ import os
 
 import click
 
-from ....utils import read_file
+from ....fs import read_file
 from ...manifest_utils import Manifest
 from ...testing import process_checks_option
 from ...utils import complete_valid_checks, get_assets_from_manifest, get_manifest_file
