+++ datadog_checks_dev/datadog_checks/dev/utils.py
@@ -5,19 +5,13 @@
 Utilities functions abstracting common operations, specially designed to be used
 by Integrations within tests.
 """
-import csv
-import inspect
 import os
 import platform
 import socket
 from contextlib import closing, contextmanager
-from io import open
-
-import yaml
 from six.moves.urllib.request import urlopen
 
 from .ci import running_on_ci, running_on_gh_actions, running_on_windows_ci  # noqa: F401
-from .fs import basepath, file_exists, get_parent_dir, path_join, read_file
 
 __platform = platform.system()
 ON_MACOS = os.name == 'mac' or __platform == 'Darwin'
@@ -42,69 +36,8 @@ def ensure_unicode(s):
     return s
 
 
-def download_file(url, fname):
-    req = urlopen(url)
-    with open(fname, 'wb') as f:
-        while True:
-            chunk = req.read(16384)
-            if not chunk:
-                break
-            f.write(chunk)
-            f.flush()
-
-
 def get_next(obj):
     return next(iter(obj))
-
-
-def load_jmx_config():
-    # Only called in tests of a check, so just go back one frame
-    root = find_check_root(depth=1)
-
-    check = basepath(root)
-    example_config_path = path_join(root, 'datadog_checks', check, 'data', 'conf.yaml.example')
-    metrics_config_path = path_join(root, 'datadog_checks', check, 'data', 'metrics.yaml')
-
-    example_config = yaml.safe_load(read_file(example_config_path))
-    metrics_config = yaml.safe_load(read_file(metrics_config_path))
-
-    if example_config['init_config'] is None:
-        example_config['init_config'] = {}
-
-    # Avoid having to potentially mount multiple files by putting the default metrics
-    # in the user-defined metric location.
-    example_config['init_config']['conf'] = metrics_config['jmx_metrics']
-
-    return example_config
-
-
-def find_check_root(depth=0):
-    # Account for this call
-    depth += 1
-
-    frame = inspect.currentframe()
-    for _ in range(depth):
-        frame = frame.f_back
-
-    root = get_parent_dir(frame.f_code.co_filename)
-    while True:
-        if file_exists(path_join(root, 'setup.py')):
-            break
-
-        new_root = os.path.dirname(root)
-        if new_root == root:
-            raise OSError('No check found')
-
-        root = new_root
-
-    return root
-
-
-def get_current_check_name(depth=0):
-    # Account for this call
-    depth += 1
-
-    return os.path.basename(find_check_root(depth))
 
 
 @contextmanager
@@ -126,14 +59,3 @@ def get_ip():
         # doesn't even have to be reachable
         s.connect(('10.255.255.255', 1))
         return s.getsockname()[0]
-
-
-def get_metadata_metrics():
-    # Only called in tests of a check, so just go back one frame
-    root = find_check_root(depth=1)
-    metadata_path = os.path.join(root, 'metadata.csv')
-    metrics = {}
-    with open(metadata_path) as f:
-        for row in csv.DictReader(f):
-            metrics[row['metric_name']] = row
-    return metrics
